requirements.txt
streamlit==1.8.1
boto3==1.21.0
pandas==1.4.1
numpy==1.22.2
pyyaml==6.0
joblib==1.1.1
pytest==8.2.0
protobuf==3.20.3
altair==4.1.0
scikit-learn==1.3.2
